backend/requirements.txt
# LeadPulse — Python Dependencies
# Install: pip install -r requirements.txt

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0          # ASGI server with websocket + reload support

# ── Supabase ──────────────────────────────────────────────────────────────────
supabase==2.4.2

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT encoding/decoding
passlib[bcrypt]==1.7.4             # Password hashing

# ── HTTP Client (Graph API calls) ─────────────────────────────────────────────
httpx==0.27.0

# ── Config ────────────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx                              # also used as test client
